[Dy2St] Remove `utils_helper`
