+++ python/paddle/jit/api.py
@@ -1889,6 +1889,21 @@ class TracedLayer:
             )
 
 
+def set_dynamic_shape(variable, shape_list):
+    if paddle.base.dygraph.base.in_to_static_mode():
+        if isinstance(variable, paddle.base.framework.Variable):
+            variable.desc.set_shape(shape_list)
+        elif isinstance(variable, paddle.pir.Value):
+            variable.set_shape(shape_list)
+        else:
+            raise TypeError(
+                "In to_static mode, variable must be a Variable or Value"
+            )
+    else:
+        # in dygraph mode, dynamic shape is not needed, just do nothing.
+        return
+
+
 def get_ast_static_function(function):
     if isinstance(function, SymbolicStaticFunction):
         if function._class_instance:
