+++ python/paddle/jit/dy2static/transformers/call_transformer.py
@@ -12,11 +12,12 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from paddle.jit.dy2static.utils import ast_to_source_code, is_paddle_api
+from paddle.jit.dy2static.utils import ast_to_source_code
 from paddle.utils import gast
 
 from ..utils import is_builtin
 from .base import BaseTransformer
+from .utils import is_paddle_api
 
 PDB_SET = "pdb.set_trace"
